♻️ refactor: fix module imports (#623)
