+++ src/engines/searx.rs
@@ -8,8 +8,8 @@ use scraper::Html;
 use std::collections::HashMap;
 
 use super::search_result_parser::SearchResultParser;
-use crate::models::aggregation_models::SearchResult;
-use crate::models::engine_models::{EngineError, SearchEngine};
+use crate::models::aggregation::SearchResult;
+use crate::models::engine::{EngineError, SearchEngine};
 use error_stack::{Report, Result, ResultExt};
 
 /// A new Searx engine type defined in-order to implement the `SearchEngine` trait which allows to
