+++ src/engines/wikipedia.rs
@@ -7,13 +7,13 @@ use reqwest::header::HeaderMap;
 use reqwest::Client;
 use scraper::Html;
 
-use crate::models::aggregation_models::SearchResult;
+use crate::models::aggregation::SearchResult;
 
-use crate::models::engine_models::{EngineError, SearchEngine};
+use crate::models::engine::{EngineError, SearchEngine};
 
 use error_stack::{Report, Result, ResultExt};
 
-use super::common::build_query;
+use super::build_query;
 use super::search_result_parser::SearchResultParser;
 
 /// A new Wikipedia engine type defined in-order to implement the `SearchEngine` trait which allows to
