+++ src/templates/views/about.rs
@@ -140,7 +140,7 @@ pub fn about(colorscheme: &str, theme: &str, animation: &Option<String>) -> Mark
                         }
                         div class="feature-card-body" {
                             p {
-                                "Websurfx comes with 9 built-in color themes and supports creating custom themes effortlessly."
+                                "Websurfx comes with 15 built-in color themes and supports creating custom themes effortlessly."
                             }
                         }
                     }
